token added

+++ backend/authentication/internal/adapters/app/api/models/AuthResBody.go
@@ -1,7 +1,7 @@
 package models
 
 type AuthResBody struct{
-	Id int64 `json:"id" xml:"id" form:"id"`
+	Id int `json:"id" xml:"id" form:"id"`
 	Email string `json:"email" xml:"email" form:"email"`
 	Token string `json:"token" xml:"token" form:"token"`
 }
